+++ submit_cgap/tests/test_submission.py
@@ -21,6 +21,7 @@ from ..submission import (
     resolve_server, resume_uploads, show_section, submit_any_ingestion,
     upload_file_to_uuid, upload_item_data, PROGRESS_CHECK_INTERVAL,
     get_s3_encrypt_key_id, get_s3_encrypt_key_id_from_health_page, running_on_windows_native,
+    search_for_file, UploadMessageWrapper, upload_extra_files,
 )
 from ..utils import FakeResponse, script_catch_errors, ERROR_HERALD
 
@@ -71,6 +72,8 @@ SOME_UPLOAD_CREDENTIALS = {
     'upload_url': SOME_UPLOAD_URL,
 }
 
+SOME_FILE_METADATA = {"upload_credentials": SOME_UPLOAD_CREDENTIALS}
+
 SOME_S3_ENCRYPT_KEY_ID = 'some/encrypt/key'
 
 SOME_EXTENDED_UPLOAD_CREDENTIALS = {
@@ -81,11 +84,7 @@ SOME_EXTENDED_UPLOAD_CREDENTIALS = {
     's3_encrypt_key_id': SOME_S3_ENCRYPT_KEY_ID,
 }
 
-SOME_UPLOAD_CREDENTIALS_RESULT = {
-    '@graph': [
-        {'upload_credentials': SOME_UPLOAD_CREDENTIALS}
-    ]
-}
+SOME_UPLOAD_CREDENTIALS_RESULT = {'@graph': [SOME_FILE_METADATA]}
 
 SOME_UPLOAD_INFO = [
     {'uuid': '1234', 'filename': 'f1.fastq.gz'},
@@ -115,6 +114,17 @@ SOME_ENVIRON_WITH_CREDS = {
     'AWS_ACCESS_KEY_ID': 'some-access-key',
     'AWS_SECRET_ACCESS_KEY': 'some-secret',
     'AWS_SECURITY_TOKEN': 'some-session-token',
+}
+
+ANOTHER_FILE_NAME = "another_file"
+
+SOME_EXTRA_FILE_CREDENTIALS = [
+    {"filename": SOME_FILENAME, "upload_credentials": SOME_ENVIRON_WITH_CREDS},
+    {"filename": ANOTHER_FILE_NAME, "upload_credentials": SOME_ENVIRON_WITH_CREDS},
+]
+
+SOME_FILE_METADATA_WITH_EXTRA_FILE_CREDENTIALS = {
+    "extra_files_creds": SOME_EXTRA_FILE_CREDENTIALS
 }
 
 
@@ -825,7 +835,8 @@ def test_upload_file_to_uuid():
 
     with mock.patch("dcicutils.ff_utils.patch_metadata", return_value=SOME_UPLOAD_CREDENTIALS_RESULT):
         with mock.patch.object(submission_module, "execute_prearranged_upload") as mocked_upload:
-            upload_file_to_uuid(filename=SOME_FILENAME, uuid=SOME_UUID, auth=SOME_AUTH)
+            metadata = upload_file_to_uuid(filename=SOME_FILENAME, uuid=SOME_UUID, auth=SOME_AUTH)
+            assert metadata == SOME_FILE_METADATA
             mocked_upload.assert_called_with(SOME_FILENAME, auth=SOME_AUTH,
                                              upload_credentials=SOME_UPLOAD_CREDENTIALS)
 
@@ -1037,6 +1048,46 @@ def test_do_uploads(tmp_path):
                 " in folder %s: %s."
                 % (filename, folder_str + "/**", ", ".join([another_file_path, file_path]))
             ]
+
+    # Test extra files credentials found and passed to handler
+    def return_first_arg(first_arg, *args, **kwargs):
+        return first_arg
+
+    mocked_instance = mock.MagicMock()
+    mocked_instance.wrap_upload_function = mock.MagicMock(side_effect=return_first_arg)
+    mocked_upload_message_wrapper = mock.MagicMock(return_value=mocked_instance)
+#    mocked_upload_message_wrapper().wrap_upload_function = mock.MagicMock(
+#        side_effect=return_first_arg
+#    )
+    with mock.patch.object(
+        submission_module,
+        "upload_file_to_uuid",
+        return_value=SOME_FILE_METADATA_WITH_EXTRA_FILE_CREDENTIALS,
+    ) as mocked_upload_file_to_uuid:
+        with mock.patch.object(
+            submission_module, "upload_extra_files"
+        ) as mocked_upload_extra_files:
+            with mock.patch.object(
+                submission_module,
+                "UploadMessageWrapper",
+                mocked_upload_message_wrapper,
+            ):
+                with shown_output() as shown:
+                    do_uploads(
+                        upload_spec_list,
+                        auth=SOME_AUTH,
+                        folder=folder,
+                        no_query=True,
+                        subfolders=False,
+                    )
+                    mocked_upload_file_to_uuid.assert_called_once()
+                    mocked_upload_extra_files.assert_called_once_with(
+                        SOME_EXTRA_FILE_CREDENTIALS,
+                        mocked_instance,
+                        folder,
+                        SOME_AUTH,
+                        recursive=False,
+                    )
 
 
 def test_upload_item_data():
@@ -2201,3 +2252,144 @@ def test_running_on_windows_native():
         os_name, is_windows = pair
         with mock.patch.object(os, "name", os_name):
             assert running_on_windows_native() is is_windows
+
+
+@pytest.mark.parametrize(
+    "directory,file_name,recursive,glob_results,expected_file_path,expected_msg",
+    [
+        ("foo", "bar", False, [], "foo/bar", False),
+        ("foo", "bar", True, [], "foo/bar", False),
+        ("foo", "bar", False, ["foo/bar"], "foo/bar", False),
+        ("foo", "bar", False, ["foo/bar", "fu/foo/bar"], None, True),
+    ]
+)
+def test_search_for_file(
+    directory, file_name, recursive, glob_results, expected_file_path, expected_msg
+):
+    """Test output file path +/- error message dependent on file search
+    via glob.
+    """
+    with mock.patch.object(
+        submission_module.glob, "glob", return_value=glob_results
+    ) as mocked_glob:
+        file_path_found, error_msg = search_for_file(directory, file_name, recursive)
+        mocked_glob.assert_called_once_with(
+            directory + "/" + file_name, recursive=recursive
+        )
+        assert file_path_found == expected_file_path
+        if expected_msg:
+            assert error_msg.startswith(
+                f"No upload attempted for file {file_name}"
+            )
+        else:
+            assert not error_msg, "Error message found when not expected"
+
+
+@pytest.mark.parametrize(
+    "no_query,cgap_selective_uploads,yes_or_no_result,error_raised,expected_result",
+    [
+        (False, True, False, None, None),
+        (False, False, False, None, None),
+        (False, True, True, None, "something"),
+        (False, True, True, True, None),
+        (True, True, False, None, "something"),
+        (True, True, False, True, None),
+    ]
+)
+def test_wrap_upload_function(
+    no_query, cgap_selective_uploads, yes_or_no_result, error_raised, expected_result
+):
+    """Test UploadMessageWrapper.wrap_upload_function creates
+    appropriate messages on given upload function.
+
+    Ensure wrapped function is indeed wrapped and returns expected
+    output when called.
+    """
+    with shown_output() as shown:
+        with mock.patch.object(
+            submission_module, "yes_or_no", return_value=yes_or_no_result
+        ) as mocked_yes_or_no:
+            with mock.patch.object(
+                submission_module,
+                "CGAP_SELECTIVE_UPLOADS",
+                cgap_selective_uploads,
+            ):
+                side_effect = None
+                if error_raised:
+                    side_effect = RuntimeError("Error occurred")
+                simple_function = mock.MagicMock(
+                    side_effect=side_effect, return_value=expected_result
+                )
+
+                uuid = "some_uuid"
+                file_name = "foo/bar"
+                input_arg = "foo"
+                function_wrapper = UploadMessageWrapper(uuid, no_query=no_query)
+                wrapped_function = function_wrapper.wrap_upload_function(
+                    simple_function, file_name
+                )
+                result = wrapped_function(input_arg, error_raised=error_raised)
+
+                expected_lines = []
+                if not no_query and cgap_selective_uploads and not yes_or_no_result:
+                    mocked_yes_or_no.assert_called_once()
+                    expected_lines.append("OK, not uploading it.")
+                    simple_function.assert_not_called()
+                else:
+                    expected_lines.append(f"Uploading {file_name} to item {uuid} ...")
+                    simple_function.assert_called_once_with(
+                        input_arg, error_raised=error_raised
+                    )
+                    if error_raised:
+                        expected_lines.append(f"RuntimeError: Error occurred")
+                    else:
+                        expected_lines.append(
+                            f"Upload of {file_name} to item {uuid} was successful."
+                        )
+                assert shown.lines == expected_lines
+                assert result == expected_result
+
+
+@pytest.mark.parametrize(
+    "credentials,files_found,expected_file_search_calls,expected_uploader_calls",
+    [
+        ([], [], 0, 0),
+        ([{"filename": "foo"}], [], 0, 0),
+        ([{"upload_credentials": {"key": "value"}}], [], 0, 0),
+        (SOME_EXTRA_FILE_CREDENTIALS, [], 2, 0),
+        (SOME_EXTRA_FILE_CREDENTIALS, [SOME_FILENAME], 2, 1),
+        (SOME_EXTRA_FILE_CREDENTIALS, [SOME_FILENAME, ANOTHER_FILE_NAME], 2, 2),
+    ]
+)
+def test_upload_extra_files(
+    credentials, files_found, expected_file_search_calls, expected_uploader_calls
+):
+    """Test extra files credentials utilized to search for and then
+    upload files.
+    """
+    folder = SOME_USER_HOMEDIR
+    recursive = True
+    auth = SOME_AUTH
+
+    def mocked_file_search(folder, extra_file_name, **kwargs):
+        if extra_file_name in files_found:
+            return os.path.join(folder, extra_file_name), None
+        else:
+            return None, "error"
+
+    with mock.patch.object(
+        submission_module, "search_for_file", side_effect=mocked_file_search
+    ) as mocked_search_for_file:
+        with mock.patch.object(
+            submission_module, "execute_prearranged_upload"
+        ) as mocked_execute_upload:
+            uploader_wrapper = UploadMessageWrapper(SOME_UUID)
+            upload_extra_files(
+                credentials,
+                uploader_wrapper,
+                folder,
+                auth,
+                recursive=recursive,
+            )
+            assert len(mocked_search_for_file.call_args_list) == expected_file_search_calls
+            assert len(mocked_execute_upload.call_args_list) == expected_uploader_calls